requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==3.0.0
Django==4.1.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
executing==1.2.0
Faker==17.5.0
filelock==3.10.0
flake8==5.0.4
flake8-bugbear==23.2.13
flake8-builtins==2.1.0
flake8-django==1.1.5
flake8-import-order==0.18.2
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-print==5.0.0
flake8-quotes==3.3.2
flake8-return==1.2.0
flake8-spellcheck==0.28.0
httpie==3.2.1
idna==3.4
ipython==8.11.0
jedi==0.18.2
markdown-it-py==2.2.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pep8-naming==0.13.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
pluggy==1.0.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.14.0
pyproject_api==1.5.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tox==4.4.6
traitlets==5.9.0
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.2.6
